[DISCO-2785] Update accuweather provider name and remove accuweather_city
