+++ merino/providers/weather/backends/accuweather.py
@@ -167,501 +167,6 @@ class WeatherDataType(Enum):
 
 
 class AccuweatherBackend:
-    """Backend that connects to the AccuWeather API."""
-
-    api_key: str
-    cache: CacheAdapter
-    cached_location_key_ttl_sec: int
-    cached_current_condition_ttl_sec: int
-    cached_forecast_ttl_sec: int
-    metrics_client: aiodogstatsd.Client
-    url_param_api_key: str
-    url_postalcodes_path: str
-    url_postalcodes_param_query: str
-    url_current_conditions_path: str
-    url_forecasts_path: str
-    url_location_key_placeholder: str
-    http_client: AsyncClient
-
-    def __init__(
-        self,
-        api_key: str,
-        cache: CacheAdapter,
-        cached_location_key_ttl_sec: int,
-        cached_current_condition_ttl_sec: int,
-        cached_forecast_ttl_sec: int,
-        metrics_client: aiodogstatsd.Client,
-        http_client: AsyncClient,
-        url_param_api_key: str,
-        url_postalcodes_path: str,
-        url_postalcodes_param_query: str,
-        url_current_conditions_path: str,
-        url_forecasts_path: str,
-        url_location_key_placeholder: str,
-    ) -> None:
-        """Initialize the AccuWeather backend.
-
-        Raises:
-            ValueError: If API key or URL parameters are None or empty.
-        """
-        if not api_key:
-            raise ValueError("AccuWeather API key not specified")
-
-        if (
-            not url_param_api_key
-            or not url_postalcodes_path
-            or not url_postalcodes_param_query
-            or not url_current_conditions_path
-            or not url_forecasts_path
-            or not url_location_key_placeholder
-        ):
-            raise ValueError("One or more AccuWeather API URL parameters are undefined")
-
-        self.api_key = api_key
-        self.cache = cache
-        # This registration is lazy (i.e. no interaction with Redis) and infallible.
-        self.cache.register_script(SCRIPT_ID, LUA_SCRIPT_CACHE_BULK_FETCH)
-        self.cached_location_key_ttl_sec = cached_location_key_ttl_sec
-        self.cached_current_condition_ttl_sec = cached_current_condition_ttl_sec
-        self.cached_forecast_ttl_sec = cached_forecast_ttl_sec
-        self.metrics_client = metrics_client
-        self.http_client = http_client
-        self.url_param_api_key = url_param_api_key
-        self.url_postalcodes_path = url_postalcodes_path
-        self.url_postalcodes_param_query = url_postalcodes_param_query
-        self.url_current_conditions_path = url_current_conditions_path
-        self.url_forecasts_path = url_forecasts_path
-        self.url_location_key_placeholder = url_location_key_placeholder
-
-    def cache_key_for_accuweather_request(
-        self, url: str, query_params: dict[str, str] = {}
-    ) -> str:
-        """Get the cache key for the accuweather request.
-        Also ensure that the API key is stripped out.
-        """
-        min_params_for_cache_key = 1 if query_params.get(self.url_param_api_key) else 0
-        if len(query_params) > min_params_for_cache_key:
-            hasher = hashlib.blake2s()
-            for key, value in sorted(query_params.items()):
-                if key != self.url_param_api_key:
-                    hasher.update(key.encode("utf-8") + value.encode("utf-8"))
-            extra_identifiers = hasher.hexdigest()
-
-            return f"{self.__class__.__name__}:v3:{url}:{extra_identifiers}"
-
-        return f"{self.__class__.__name__}:v3:{url}"
-
-    @functools.cache
-    def cache_key_template(self, dt: WeatherDataType) -> str:
-        """Get the cache key template for weather data."""
-        query_params: dict[str, str] = {self.url_param_api_key: self.api_key}
-        match dt:
-            case WeatherDataType.CURRENT_CONDITIONS:
-                return self.cache_key_for_accuweather_request(
-                    self.url_current_conditions_path,
-                    query_params=query_params,
-                )
-            case WeatherDataType.FORECAST:  # pragma: no cover
-                return self.cache_key_for_accuweather_request(
-                    self.url_forecasts_path,
-                    query_params=query_params,
-                )
-
-    async def get_request(
-        self,
-        url_path: str,
-        params: dict[str, str],
-        process_api_response: Callable[[Any], dict[str, Any] | None],
-        cache_ttl_sec: int,
-    ) -> dict[str, Any] | None:
-        """Get API response. Attempt to get it from cache first,
-        then actually make the call if there's a cache miss.
-        """
-        cache_key = self.cache_key_for_accuweather_request(url_path, params)
-        response_dict: dict[str, Any] | None
-
-        # The top level path in the URL gives us a good enough idea of what type of request
-        # we are calling from here.
-        request_type: str = url_path.strip("/").split("/", 1)[0]
-
-        with self.metrics_client.timeit(f"accuweather.request.{request_type}.get"):
-            response: Response = await self.http_client.get(url_path, params=params)
-            response.raise_for_status()
-
-        if (response_dict := process_api_response(response.json())) is None:
-            self.metrics_client.increment(
-                f"accuweather.request.{request_type}.processor.error"
-            )
-            return None
-
-        response_expiry: str = response.headers.get("Expires")
-        try:
-            cached_request_ttl = await self.store_request_into_cache(
-                cache_key, response_dict, response_expiry, cache_ttl_sec
-            )
-            # add the ttl of the just cached request to the response dict we return
-            response_dict["cached_request_ttl"] = cached_request_ttl
-        except (CacheAdapterError, ValueError) as exc:
-            logger.error(f"Error with storing Accuweather to cache: {exc}")
-            error_type = (
-                "set_error" if isinstance(exc, CacheAdapterError) else "ttl_date_error"
-            )
-            self.metrics_client.increment(f"accuweather.cache.store.{error_type}")
-            raise AccuweatherError(
-                "Something went wrong with storing to cache. Did not update cache."
-            )
-
-        return response_dict
-
-    async def store_request_into_cache(
-        self,
-        cache_key: str,
-        response_dict: dict[str, Any],
-        response_expiry: str,
-        cache_ttl_sec: int,
-    ) -> int:
-        """Store the request into cache. Also ensures that the cache ttl is
-        at least `cached_ttl_sec`. Returns the cached request's ttl in seconds.
-        """
-        with self.metrics_client.timeit("accuweather.cache.store"):
-            expiry_delta: datetime.timedelta = parser.parse(
-                response_expiry
-            ) - datetime.datetime.now(datetime.timezone.utc)
-            cache_ttl: datetime.timedelta = max(
-                expiry_delta, datetime.timedelta(seconds=cache_ttl_sec)
-            )
-            cache_value = json.dumps(response_dict).encode("utf-8")
-            await self.cache.set(cache_key, cache_value, ttl=cache_ttl)
-
-        return cache_ttl.seconds
-
-    def emit_cache_fetch_metrics(self, cached_data: list[bytes | None]) -> None:
-        """Emit cache fetch metrics.
-
-        Params:
-            - `cached_data` {list[bytes]} A list of bytes for location_key,
-              current_condition, forecast
-        """
-        location, current, forecast = False, False, False
-        match cached_data:
-            case []:
-                pass
-            # the last variable is ttl but is omitted here since we don't need to use but need
-            # it to satisfy this match case
-            case [location_cached, current_cached, forecast_cached, _]:
-                location, current, forecast = (
-                    location_cached is not None,
-                    current_cached is not None,
-                    forecast_cached is not None,
-                )
-            case _:  # pragma: no cover
-                pass
-
-        self.metrics_client.increment(
-            "accuweather.cache.hit.locations"
-            if location
-            else "accuweather.cache.fetch.miss.locations"
-        )
-        self.metrics_client.increment(
-            "accuweather.cache.hit.currentconditions"
-            if current
-            else "accuweather.cache.fetch.miss.currentconditions"
-        )
-        self.metrics_client.increment(
-            "accuweather.cache.hit.forecasts"
-            if forecast
-            else "accuweather.cache.fetch.miss.forecasts"
-        )
-
-    def parse_cached_data(self, cached_data: list[bytes | None]) -> WeatherData:
-        """Parse the weather data from cache.
-
-        Upon parsing errors, it will return the successfully parsed data thus far.
-
-        Params:
-            - `cached_data` {list[bytes]} A list of bytes for location_key,
-              current_conditions, forecast, ttl
-        """
-        if len(cached_data) == 0:
-            return WeatherData()
-
-        location_cached, current_cached, forecast_cached, ttl_cached = cached_data
-
-        location: AccuweatherLocation | None = None
-        current_conditions: CurrentConditions | None = None
-        forecast: Forecast | None = None
-        ttl: int | None = None
-
-        try:
-            if location_cached is not None:
-                location = AccuweatherLocation.model_validate_json(location_cached)
-            if current_cached is not None:
-                current_conditions = CurrentConditions.model_validate_json(
-                    current_cached
-                )
-            if forecast_cached is not None:
-                forecast = Forecast.model_validate_json(forecast_cached)
-            if ttl_cached is not None:
-                # redis returns the TTL value as an integer, however, we are explicitly casting
-                # the value returned from the cache since it's received as bytes as the method
-                # argument. This satisfies the type checker.
-                ttl = cast(int, ttl_cached)
-        except ValidationError as exc:
-            logger.error(f"Failed to load weather report data from Redis: {exc}")
-            self.metrics_client.increment("accuweather.cache.data.error")
-
-        return WeatherData(location, current_conditions, forecast, ttl)
-
-    def get_location_key_query_params(self, postal_code: str) -> dict[str, str]:
-        """Get the query parameters for the location key for a given postal code."""
-        return {
-            self.url_param_api_key: self.api_key,
-            self.url_postalcodes_param_query: postal_code,
-        }
-
-    async def get_weather_report(
-        self, geolocation: Location, location_key: str | None = None
-    ) -> WeatherReport | None:
-        """Get weather information from AccuWeather.
-
-        Firstly, it will look up the Redis cache for the location key, current condition,
-        and forecast. If all of them are found in the cache, then return them without
-        requesting those from AccuWeather. Otherwise, it will issue API requests to
-        AccuWeather for the missing data. Lastly, the API responses are stored in the
-        cache for future uses.
-
-        Note:
-            - To avoid making excessive API requests to Accuweather in the event of
-              "Cache Avalanche", it will *not* call AccuWeather for weather reports upon any
-              cache errors such as timeouts or connection issues to Redis
-
-        Raises:
-            AccuweatherError: Failed request or 4xx and 5xx response from AccuWeather.
-        """
-        country: str | None = geolocation.country
-        postal_code: str | None = geolocation.postal_code
-        if not country or not postal_code:
-            raise AccuweatherError("Country and/or postal code unknown")
-
-        cache_key: str = self.cache_key_for_accuweather_request(
-            self.url_postalcodes_path.format(country_code=country),
-            query_params=self.get_location_key_query_params(postal_code),
-        )
-        # Look up for all the weather data from the cache.
-        try:
-            with self.metrics_client.timeit("accuweather.cache.fetch"):
-                cached_data: list[bytes | None] = await self.cache.run_script(
-                    sid=SCRIPT_ID,
-                    keys=[cache_key],
-                    # The order matters below. See `LUA_SCRIPT_CACHE_BULK_FETCH` for details.
-                    args=[
-                        self.cache_key_template(WeatherDataType.CURRENT_CONDITIONS),
-                        self.cache_key_template(WeatherDataType.FORECAST),
-                        self.url_location_key_placeholder,
-                    ],
-                )
-        except CacheAdapterError as exc:
-            logger.error(f"Failed to fetch weather report from Redis: {exc}")
-            self.metrics_client.increment("accuweather.cache.fetch.error")
-            return None
-
-        self.emit_cache_fetch_metrics(cached_data)
-
-        cached_report = self.parse_cached_data(cached_data)
-        return await self.make_weather_report(cached_report, country, postal_code)
-
-    async def make_weather_report(
-        self,
-        cached_report: WeatherData,
-        country: str,
-        postal_code: str,
-    ) -> WeatherReport | None:
-        """Make a `WeatherReport` either using the cached data or fetching from AccuWeather.
-
-        Raises:
-            AccuWeatherError: Failed request or 4xx and 5xx response from AccuWeather.
-        """
-
-        async def as_awaitable(val: Any) -> Any:
-            """Wrap a non-awaitable value into a coroutine and resolve it right away."""
-            return val
-
-        location, current_conditions, forecast, ttl = cached_report
-
-        # validate the integer value of the TTL received from the cache. log a warning if there was
-        # an error getting the ttl from the cache
-        if ttl_error_message := get_ttl_error_message(ttl):
-            logger.warning(ttl_error_message)
-
-        if location and current_conditions and forecast and ttl:
-            # Return the weather report with the values returned from the cache.
-            # `ttl` here could be the cached ttl value or None.
-            return WeatherReport(
-                city_name=location.localized_name,
-                current_conditions=current_conditions,
-                forecast=forecast,
-                ttl=ttl,
-            )
-
-        # The cached report is incomplete, now fetching from AccuWeather.
-        if location is None:
-            if (location := await self.get_location(country, postal_code)) is None:
-                return None
-
-        try:
-            async with asyncio.TaskGroup() as tg:
-                task_current = (
-                    tg.create_task(self.get_current_conditions(location.key))
-                    if current_conditions is None
-                    else as_awaitable(
-                        CurrentConditionsWithTTL(
-                            current_conditions=current_conditions,
-                            ttl=self.cached_current_condition_ttl_sec,
-                        )
-                    )
-                )
-                task_forecast = (
-                    tg.create_task(self.get_forecast(location.key))
-                    if forecast is None
-                    else as_awaitable(
-                        ForecastWithTTL(
-                            forecast=forecast, ttl=self.cached_forecast_ttl_sec
-                        )
-                    )
-                )
-        except ExceptionGroup as e:
-            raise AccuweatherError(f"Failed to fetch weather report: {e.exceptions}")
-
-        if (current_conditions_response := await task_current) is not None and (
-            forecast_response := await task_forecast
-        ):
-            current_conditions, current_conditions_ttl = current_conditions_response
-            forecast, forecast_ttl = forecast_response
-            weather_report_ttl = min(current_conditions_ttl, forecast_ttl)
-
-            return (
-                WeatherReport(
-                    city_name=location.localized_name,
-                    current_conditions=current_conditions,
-                    forecast=forecast,
-                    ttl=weather_report_ttl,
-                )
-                if current_conditions is not None and forecast is not None
-                else None
-            )
-        else:
-            return None
-
-    async def get_location(
-        self, country: str, postal_code: str
-    ) -> AccuweatherLocation | None:
-        """Return location data for a specific country and postal code or None if
-        location data is not found.
-
-        Raises:
-            AccuweatherError: Failed request or 4xx and 5xx response from AccuWeather.
-        Reference:
-            https://developer.accuweather.com/accuweather-locations-api/apis/get/locations/v1/postalcodes/{countryCode}/search
-        """
-        try:
-            response: dict[str, Any] | None = await self.get_request(
-                self.url_postalcodes_path.format(country_code=country),
-                params=self.get_location_key_query_params(postal_code),
-                process_api_response=process_location_response,
-                cache_ttl_sec=self.cached_location_key_ttl_sec,
-            )
-        except HTTPError as error:
-            raise AccuweatherError("Unexpected location response") from error
-
-        return AccuweatherLocation(**response) if response else None
-
-    async def get_current_conditions(
-        self, location_key: str
-    ) -> CurrentConditionsWithTTL | None:
-        """Return current conditions data for a specific location or None if current
-        conditions data is not found.
-
-        Raises:
-            AccuweatherError: Failed request or 4xx and 5xx response from AccuWeather.
-        Reference:
-            https://developer.accuweather.com/accuweather-current-conditions-api/apis/get/currentconditions/v1/{locationKey}
-        """
-        try:
-            response: dict[str, Any] | None = await self.get_request(
-                self.url_current_conditions_path.format(location_key=location_key),
-                params={
-                    self.url_param_api_key: self.api_key,
-                },
-                process_api_response=process_current_condition_response,
-                cache_ttl_sec=self.cached_current_condition_ttl_sec,
-            )
-        except HTTPError as error:
-            raise AccuweatherError("Unexpected current conditions response") from error
-
-        return (
-            CurrentConditionsWithTTL(
-                current_conditions=CurrentConditions(
-                    url=response["url"],
-                    summary=response["summary"],
-                    icon_id=response["icon_id"],
-                    temperature=Temperature(**response["temperature"]),
-                ),
-                ttl=response["cached_request_ttl"],
-            )
-            if response
-            else None
-        )
-
-    async def get_forecast(self, location_key: str) -> ForecastWithTTL | None:
-        """Return daily forecast data for a specific location or None if daily
-        forecast data is not found.
-
-        Raises:
-            AccuweatherError: Failed request or 4xx and 5xx response from AccuWeather.
-        Reference:
-            https://developer.accuweather.com/accuweather-forecast-api/apis/get/forecasts/v1/daily/1day/{locationKey}
-        """
-        try:
-            response: dict[str, Any] | None = await self.get_request(
-                self.url_forecasts_path.format(location_key=location_key),
-                params={
-                    self.url_param_api_key: self.api_key,
-                },
-                process_api_response=process_forecast_response,
-                cache_ttl_sec=self.cached_forecast_ttl_sec,
-            )
-        except HTTPError as error:
-            raise AccuweatherError("Unexpected forecast response") from error
-
-        return (
-            ForecastWithTTL(
-                forecast=Forecast(
-                    url=response["url"],
-                    summary=response["summary"],
-                    high=Temperature(**response["high"]),
-                    low=Temperature(**response["low"]),
-                ),
-                ttl=response["cached_request_ttl"],
-            )
-            if response
-            else None
-        )
-
-    # NOTE: This method is not being used for this class / backend. This only here to satisfy mypy
-    # type checker. This will be removed once we cut over to the AccuweatherCityBackend
-    async def get_location_completion(
-        self, geolocation: Location, search_term: str
-    ) -> list[LocationCompletion] | None:
-        """Fetch a list of locations from the Accuweather API given a search term and location."""
-        ...
-
-    async def shutdown(self) -> None:
-        """Close out the cache during shutdown."""
-        await self.http_client.aclose()
-        await self.cache.close()
-
-
-class AccuweatherCityBackend:
     """Backend that connects to the AccuWeather API using City to find location key."""
 
     api_key: str
@@ -954,7 +459,7 @@ class AccuweatherCityBackend:
         """
         # Look up for all the weather data from the cache.
         try:
-            with self.metrics_client.timeit("accuweather.cache.fetch-via-location-key"):
+            with self.metrics_client.timeit("accuweather.cache.fetch"):
                 cached_data: list[bytes | None] = await self.cache.run_script(
                     sid=SCRIPT_LOCATION_KEY_ID,
                     keys=[],
@@ -973,9 +478,7 @@ class AccuweatherCityBackend:
                 )
         except CacheAdapterError as exc:
             logger.error(f"Failed to fetch weather report from Redis: {exc}")
-            self.metrics_client.increment(
-                "accuweather.cache.fetch-via-location-key.error"
-            )
+            self.metrics_client.increment("accuweather.cache.fetch.error")
             return None
 
         self.emit_cache_fetch_metrics(cached_data, skip_location_key=True)
